refactor: rename correctColor
